requirements.txt
opencv-python
numpy
Pillow
scikit-image
